feat: update docker compose, improve like unlike logic

+++ src/services/likes.service.ts
@@ -2,8 +2,11 @@ import { supabase } from "@/utils/supabaseClient";
 
 const likePost = async (postId: string) => {
 	const { error } = await supabase.from("likes").insert({ post_id: postId });
-	if (error) {
-		throw new Error(error.message);
+	const { error: incrementError } = await supabase.rpc("increment_like_count", {
+		row_id: postId,
+	});
+	if (error || incrementError) {
+		throw new Error(error?.message || incrementError?.message);
 	}
 };
 
@@ -11,10 +14,15 @@ const unlikePost = async (postId: string) => {
 	const { data, error } = await supabase
 		.from("likes")
 		.delete()
-		.eq("post_id", postId);
+		.eq("post_id", postId)
+		.eq("user_id", (await supabase.auth.getUser()).data.user?.id);
 
-	if (error) {
-		throw new Error(error.message);
+	const { error: decrementError } = await supabase.rpc("decrement_like_count", {
+		row_id: postId,
+	});
+
+	if (error || decrementError) {
+		throw new Error(error?.message || decrementError?.message);
 	}
 
 	return data;
